Defensive check for DB upgrade to version 72
A constant for a column was renamed from 'IS_NTN' to 'IS_ONLY_NTN'. This change was incorrected backported to the v74 upgrade logic.
because the column name IS_NTN was removed.
However, in case a device was upgrade from under v63 to above v72, this causes DB duplication exception.

Bug: 431071244
Flag: EXEMPT bug fix
Test: manually run db upgrade from undr v63 to above v72 then check whethere exception happened
      Skip adding a unit test due to it is required to handle static method dependencies and XML parsing during DatabaseHelper instantiation. This fix was validated manually instead. Test Result : b/434647889
Change-Id: I38d73d17ea90c214362e67b7ba006865390c9f57